Add files via upload
Generate random diffusion on a 2D grid. Usage:
python 2D_walk.py -a 20 -r 2 -n 10
Apply a diffusion on a 20*20 grid, with 10 random defect MOFs, each defect region is expended to a radius of 2.
the code makes 2 figures, one with no path and one with a random diffusion path through the lattice.